requirements.txt
Django>=5.1.6,<6.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.5.0
django-pandas>=0.6.7
xero-python>=4.0.0
pandas>=2.0.0
pandas-gbq>=0.19.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-cloud-bigquery>=3.38.0
pydata-google-auth>=1.9.0
pytz>=2023.3
requests>=2.31.0
apscheduler>=3.10.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
whitenoise>=6.0.0
django-cors-headers>=4.0.0
gunicorn>=21.0.0
uvicorn>=0.30.0
channels>=4.0.0
yfinance>=0.2.0
# AI Agent MCP Skills Engine
TM1py>=2.0.0
anthropic>=0.40.0
openai>=1.50.0
voyageai>=0.3.0
duckduckgo-search>=7.0.0
httpx>=0.27.0
